+++ packages/wasm-dpp/test/integration/identity/validation/validatePublicKeysFactory.spec.js
@@ -2,45 +2,45 @@ const { getRE2Class } = require('@dashevo/wasm-re2');
 
 const crypto = require('crypto');
 
-const createAjv = require('../../../../lib/ajv/createAjv');
+const createAjv = require('@dashevo/dpp/lib/ajv/createAjv');
 
 const JsonSchemaValidator = require(
-  '../../../../lib/validation/JsonSchemaValidator',
+  '@dashevo/dpp/lib/validation/JsonSchemaValidator',
 );
 
 const validatePublicKeysFactory = require(
-  '../../../../lib/identity/validation/validatePublicKeysFactory',
+  '@dashevo/dpp/lib/identity/validation/validatePublicKeysFactory',
 );
 
-const getIdentityFixture = require('../../../../lib/test/fixtures/getIdentityFixture');
+const getIdentityFixture = require('@dashevo/dpp/lib/test/fixtures/getIdentityFixture');
 
 const {
   expectValidationError,
   expectJsonSchemaError,
-} = require('../../../../lib/test/expect/expectError');
+} = require('@dashevo/dpp/lib/test/expect/expectError');
 
 const DuplicatedIdentityPublicKeyError = require(
-  '../../../../lib/errors/consensus/basic/identity/DuplicatedIdentityPublicKeyError',
+  '@dashevo/dpp/lib/errors/consensus/basic/identity/DuplicatedIdentityPublicKeyError',
 );
 const DuplicatedIdentityPublicKeyIdError = require(
-  '../../../../lib/errors/consensus/basic/identity/DuplicatedIdentityPublicKeyIdError',
+  '@dashevo/dpp/lib/errors/consensus/basic/identity/DuplicatedIdentityPublicKeyIdError',
 );
 
 const InvalidIdentityPublicKeyDataError = require(
-  '../../../../lib/errors/consensus/basic/identity/InvalidIdentityPublicKeyDataError',
+  '@dashevo/dpp/lib/errors/consensus/basic/identity/InvalidIdentityPublicKeyDataError',
 );
 
 const InvalidIdentityPublicKeySecurityLevelError = require(
-  '../../../../lib/errors/consensus/basic/identity/InvalidIdentityPublicKeySecurityLevelError',
+  '@dashevo/dpp/lib/errors/consensus/basic/identity/InvalidIdentityPublicKeySecurityLevelError',
 );
 
 const IdentityPublicKey = require(
-  '../../../../lib/identity/IdentityPublicKey',
+  '@dashevo/dpp/lib/identity/IdentityPublicKey',
 );
-const BlsSignatures = require('../../../../lib/bls/bls');
+const BlsSignatures = require('@dashevo/dpp/lib/bls/bls');
 
-const identityPublicKeySchema = require('../../../../schema/identity/publicKey.json');
-const stateTransitionPublicKeySchema = require('../../../../schema/identity/stateTransition/publicKey.json');
+const identityPublicKeySchema = require('@dashevo/dpp/schema/identity/publicKey.json');
+const stateTransitionPublicKeySchema = require('@dashevo/dpp/schema/identity/stateTransition/publicKey.json');
 
 describe('validatePublicKeysFactory', () => {
   let rawPublicKeys;